requirements.txt
openai
praw
python-binance
TA-Lib
PyYAML
pandas
numpy
matplotlib